Reworked Structure and added some Funktions to the homescreen
