+++ lib/home.dart
@@ -16,21 +16,73 @@ class _State extends State<HomeScreen> {
   }
 
   @override
-  Widget build(context) {
-    return FutureBuilder<int>(
-        future: onlinePlayer(),
-        builder: (context, AsyncSnapshot<int> snapshot) {
-          if (snapshot.hasData) {
-            return Container(
-              color: Colors.grey[300],
-              child:Center(child: Text(snapshot.data.toString())));
-          } else {
-            return Center(
-              child: Container(
-                child: CircularProgressIndicator(),
+  Widget build(BuildContext context) {
+    return Scaffold(
+      drawer: Drawer(
+        child: Container(
+          child: ListView(
+            children: [
+              Container(
+                child: Center(
+                  child: new Container(
+                      width: 200,
+                      height: 200,
+                      decoration: new BoxDecoration(
+                          shape: BoxShape.circle,
+                          image: new DecorationImage(
+                              fit: BoxFit.cover,
+                              image: new NetworkImage(
+                                  "https://raw.githubusercontent.com/cryptic-game/graphics/master/app_icon/cryptic_mobile/1024x1024.png")))),
+                ),
+                color: Color(0xff121212),
               ),
-            );
-          }
-        });
+              ListTile(
+                  tileColor: Colors.grey,
+                  title: Text('Logout'),
+                  onTap: () {
+                    Navigator.of(context).pop();
+                    Navigator.of(context).pop();
+                  }),
+              FutureBuilder<int>(
+                  future: onlinePlayer(),
+                  builder: (context, AsyncSnapshot<int> snapshot) {
+                    if (snapshot.hasData) {
+                      return ListTile(
+                        tileColor: Colors.grey,
+                        title: Text('Spieler online'),
+                        onTap: () {
+                          return AlertDialog(
+                            title: Center(
+                                child: Text(
+                                    'Es sind momentan $snapshot.data.toString() online')),
+                            backgroundColor: Colors.grey[300],
+                            actions: <Widget>[
+                              FlatButton(
+                                child: Text('Ok'),
+                                onPressed: () {
+                                  Navigator.of(context).pop();
+                                },
+                              ),
+                            ],
+                          );
+                        },
+                      );
+                    } else {
+                      return Center(
+                        child: Container(
+                          child: CircularProgressIndicator(),
+                        ),
+                      );
+                    }
+                  })
+            ],
+          ),
+          color: Color(0xff121212),
+        ),
+      ),
+      appBar: AppBar(
+        backgroundColor: Colors.lime,
+      ),
+    );
   }
 }
