+++ pubspec.yaml
@@ -47,6 +47,7 @@ flutter:
 
   assets:
     - assets/hintergrund.jpeg
+    - assets/87x87.png
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
